# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/doctor": {
            "get": {
                "tags": [
                    "Doctor API CRUD"
                ],
                "summary": "Get all Doctors",
                "description": "Retrieve all Doctors",
                "operationId": "doctor_index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctor API CRUD"
                ],
                "summary": "Create a new Doctor",
                "description": "Add a new Doctor",
                "operationId": "doctor_store",
                "requestBody": {
                    "description": "Doctor save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "password",
                                    "email",
                                    "contact",
                                    "photo",
                                    "about",
                                    "education",
                                    "practice",
                                    "residency"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "photo": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "education": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "practice": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "residency": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/v1/doctor/{id}": {
            "get": {
                "tags": [
                    "Doctor API CRUD"
                ],
                "summary": "Get a single Doctor by ID",
                "description": "Retrieve a single Doctor by its ID",
                "operationId": "doctor_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Doctor to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctor API CRUD"
                ],
                "summary": "Update a Doctor by ID",
                "description": "Update a specific Doctor by its ID",
                "operationId": "doctor_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Doctor to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Doctor save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "contact",
                                    "photo",
                                    "about",
                                    "education",
                                    "practice",
                                    "residency"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "photo": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "education": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "practice": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "residency": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "_method": {
                                        "description": "Read-only: This property cannot be modified.",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Doctor API CRUD"
                ],
                "summary": "Delete a Doctor by ID",
                "description": "Remove a specific Doctor by its ID",
                "operationId": "doctor_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Doctor to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Doctor": {
                "title": "Doctor title",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "firstname": {
                                    "type": "string"
                                },
                                "lastname": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "contact": {
                                    "type": "string"
                                },
                                "photo": {
                                    "type": "file"
                                },
                                "about": {
                                    "type": "string"
                                },
                                "education": {
                                    "type": "string"
                                },
                                "practice": {
                                    "type": "string"
                                },
                                "residency": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": []
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            }
        }
    }
}